+++ hari_client/upload/hari_uploader.py
@@ -245,6 +245,7 @@ class HARIUploader:
 def _merge_bulk_responses(*args: models.BulkResponse) -> models.BulkResponse:
     """
     Merges multiple BulkResponse objects into one.
+    If no BulkResponse objects are provided, an empty BulkResponse object with status SUCCESS is returned.
     If only one BulkResponse object is provided, it will be returned as is.
 
     Args:
@@ -253,10 +254,14 @@ def _merge_bulk_responses(*args: models.BulkResponse) -> models.BulkResponse:
     Returns:
         models.BulkResponse: The merged BulkResponse object
     """
+    final_response = models.BulkResponse()
+
+    if len(args) == 0:
+        final_response.status = models.BulkOperationStatusEnum.SUCCESS
+        return final_response
+
     if len(args) == 1:
         return args[0]
-
-    final_response = models.BulkResponse()
 
     statuses = set()