+++ hari_client/upload/hari_uploader.py
@@ -300,9 +300,11 @@ class HARIUploader:
         media_upload_bulk_response: models.BulkResponse,
     ) -> None:
         for media in medias_to_upload:
+            if len(media.media_objects) == 0:
+                continue
             # from the endpoints we used, we know that the results items are of type
-            # models.AnnotatableCreateResponse, which contains the bulk_operation_
-            # annotatable_id.
+            # models.AnnotatableCreateResponse, which contains
+            # the bulk_operation_annotatable_id.
             filtered_upload_response = list(
                 filter(
                     lambda x: x.bulk_operation_annotatable_id
@@ -335,8 +337,8 @@ class HARIUploader:
             if len(media_object.attributes) == 0:
                 continue
             # from the endpoints we used, we know that the results items are of type
-            # models.AnnotatableCreateResponse, which contains the bulk_operation_
-            # annotatable_id.
+            # models.AnnotatableCreateResponse, which contains
+            # the bulk_operation_annotatable_id.
             filtered_upload_response = list(
                 filter(
                     lambda x: x.bulk_operation_annotatable_id
@@ -344,8 +346,6 @@ class HARIUploader:
                     media_object_upload_bulk_response.results,
                 )
             )
-            if len(media_object.attributes) == 0:
-                return
             if len(filtered_upload_response) == 0:
                 raise HARIMediaObjectUploadError(
                     f"MediaObject upload response doesn't match expectation. Couldn't find "
@@ -372,8 +372,8 @@ class HARIUploader:
             if len(media.attributes) == 0:
                 continue
             # from the endpoints we used, we know that the results items are of type
-            # models.AnnotatableCreateResponse, which contains the bulk_operation_
-            # annotatable_id.
+            # models.AnnotatableCreateResponse, which contains
+            # the bulk_operation_annotatable_id.
             filtered_upload_response = list(
                 filter(
                     lambda x: x.bulk_operation_annotatable_id
@@ -407,6 +407,7 @@ class HARIUploader:
 def _merge_bulk_responses(*args: models.BulkResponse) -> models.BulkResponse:
     """
     Merges multiple BulkResponse objects into one.
+    If no BulkResponse objects are provided, an empty BulkResponse object with status SUCCESS is returned.
     If only one BulkResponse object is provided, it will be returned as is.
 
     Args:
@@ -415,10 +416,14 @@ def _merge_bulk_responses(*args: models.BulkResponse) -> models.BulkResponse:
     Returns:
         models.BulkResponse: The merged BulkResponse object
     """
+    final_response = models.BulkResponse()
+
+    if len(args) == 0:
+        final_response.status = models.BulkOperationStatusEnum.SUCCESS
+        return final_response
+
     if len(args) == 1:
         return args[0]
-
-    final_response = models.BulkResponse()
 
     statuses = set()