+++ hari_client/models/models.py
@@ -921,6 +921,95 @@ class AttributeCreate(BaseModel):
         return values
 
 
+class Attribute(BaseModel):
+    id: str = pydantic.Field(title="ID")
+    name: str
+    question: str
+    annotatable_id: str
+    annotatable_type: DataBaseObjectType
+    subset_ids: set[str] = set()
+    attribute_type: AttributeType | None = None
+    attribute_group: AttributeGroup
+    value: typeT
+    min: typeT | None = None
+    max: typeT | None = None
+    sum: typeT | None = None
+    cant_solves: int | None = None
+    solvability: float | None = None
+    aggregate: typeT | None = None
+    modal: typeT | None = None
+    credibility: float | None = None
+    convergence: float | None = None
+    ambiguity: float | None = None
+    median: typeT | None = None
+    variance: float | None = None
+    standard_deviation: float | None = None
+    range: typing.Any | None = None
+    average_absolute_deviation: float | None = None
+    cumulated_frequency: typing.Any | None = None
+    frequency: dict[str, int] | None = None
+    question: str | None = None
+    ml_predictions: dict[str, float] | None = None
+    ml_probability_distributions: dict[str, float] | None = None
+
+
+class AttributeResponse(BaseModel):
+    id: str | None = pydantic.Field(default=None, title="Id")
+    dataset_id: str | None = pydantic.Field(default=None, title="Dataset Id")
+    timestamp: str | None = pydantic.Field(default=None, title="Timestamp")
+    archived: bool | None = pydantic.Field(default=None, title="Archived")
+    subset_ids: set[str] = pydantic.Field(default=set(), title="Subset Ids")
+    metadata_id: str | None = pydantic.Field(default=None, title="Metadata Id")
+    name: str | None = pydantic.Field(default=None, title="Name")
+    question: str | None = pydantic.Field(default=None, title="Question")
+    annotatable_id: str | None = pydantic.Field(default=None, title="Annotatable ID")
+    annotatable_type: DataBaseObjectType | None = pydantic.Field(
+        default=None, title="Annotatable Type"
+    )
+    attribute_type: AttributeType | None = pydantic.Field(
+        default=None, title="Attribute Type"
+    )
+    attribute_group: AttributeGroup | None = pydantic.Field(
+        default=None, title="Attribute Group"
+    )
+    value: typeT | None = pydantic.Field(default=None, title="Value")
+    min: typeT | None = pydantic.Field(default=None, title="Min")
+    max: typeT | None = pydantic.Field(default=None, title="Max")
+    sum: typeT | None = pydantic.Field(default=None, title="Sum")
+    cant_solves: int | None = pydantic.Field(default=None, title="Cant Solves")
+    solvability: float | None = pydantic.Field(default=None, title="Solvability")
+    aggregate: typeT | None = pydantic.Field(default=None, title="Aggregate")
+    modal: typeT | None = pydantic.Field(default=None, title="Modal")
+    credibility: float | None = pydantic.Field(default=None, title="Credibility")
+    convergence: float | None = pydantic.Field(default=None, title="Convergence")
+    ambiguity: float | None = pydantic.Field(default=None, title="Ambiguity")
+    median: typeT | None = pydantic.Field(default=None, title="Median")
+    variance: float | None = pydantic.Field(default=None, title="Variance")
+    standard_deviation: float | None = pydantic.Field(
+        default=None, title="Standard Deviation"
+    )
+    range: typing.Any | None = pydantic.Field(default=None, title="Range")
+    average_absolute_deviation: float | None = pydantic.Field(
+        default=None, title="Average Absolute Deviation"
+    )
+    cumulated_frequency: typing.Any | None = pydantic.Field(
+        default=None, title="Cumulated Frequency"
+    )
+    frequency: dict[str, int] | None = pydantic.Field(default=None, title="Frequency")
+    ml_predictions: dict[str, float] | None = pydantic.Field(
+        default=None,
+        title="ML Predictions",
+        description="These are the parameters of the posterior Dirichlet distribution",
+    )
+    ml_probability_distributions: dict[str, float] | None = pydantic.Field(
+        default=None,
+        title="ML Probability Distributions",
+        description="A point estimate for the probability associated with each category"
+        ", obtained from the full Dirichlet distribution predicted by the"
+        " model.",
+    )
+
+
 class BulkAttributeCreate(AttributeCreate):
     pass