+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## [major.minor.patch] - YYYY-MM-DD
 
+### Fixes
+
+- fixed merge_bulk_responses for the case when no responses are given [PR#20](https://github.com/quality-match/hari-client/pull/20)
+
 ### Internal
 
-- update old type-hinting conventions for is 3.11
-- fix some str/uuid inconsistencies for models
+- updated old type-hinting conventions for is 3.11 [PR#19](https://github.com/quality-match/hari-client/pull/19)
+- fixed some str/uuid inconsistencies for models [PR#19](https://github.com/quality-match/hari-client/pull/19)
 
 ## [1.0.0] - 27.09.2024