+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [major.minor.patch] - YYYY-MM-DD
 
+### Fixes
+
+- fixed merge_bulk_responses for the case when no responses are given, therefore uploading media without media_objects isn't identified as an unsuccessful upload anymore [PR#20](https://github.com/quality-match/hari-client/pull/20)
+
+### Internal
+
+- updated occurrences of old type-hinting conventions (e.g. using pipe (`|`) instead of `typing.Optional` and `typing.Union`) [PR#19](https://github.com/quality-match/hari-client/pull/19)
+- fixed some str/uuid inconsistencies for models [PR#19](https://github.com/quality-match/hari-client/pull/19)
+
 ## [1.0.0] - 27.09.2024
 
 ### New Features